requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
tensorflow==2.7
